requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
python-jose
passlib[bcrypt]
python-multipart